Refactor to utilise global constant memory

+++ spotfinder/spotfinder.cc
@@ -378,8 +378,8 @@ int main(int argc, char **argv) {
     uint32_t num_images = parser.is_used("images") ? parser.get<uint32_t>("images")
                                                    : reader.get_number_of_images();
 
-    int height = reader.image_shape()[0];
-    int width = reader.image_shape()[1];
+    ushort height = reader.image_shape()[0];
+    ushort width = reader.image_shape()[1];
     auto trusted_px_max = reader.get_trusted_range()[1];
 
     detector_geometry detector;
